refactor(poster): Restructure the files
