+++ src/HandlerSwitcher.tsx
@@ -10,8 +10,8 @@ import { default as DefaultView, LoadingView, ReloadCountdown } from './handlers
 import { AuthContext } from './contexts/AuthContext';
 import StageEffectsView from './handlers/stage-effects';
 import TimeTrailRaceView from './handlers/time-trail-race';
-import PosterGewisView from './handlers/poster-gewis';
-import PosterHubbleView from './handlers/poster-hubble';
+import PosterGewisView from './handlers/poster/gewis';
+import PosterHubbleView from './handlers/poster/hubble';
 import RoomResponsibleLegacyView from './handlers/room-responsible-legacy';
 
 export enum Handlers {
