+++ src/handlers/poster/components/ProgressBarSlider.tsx
