+++ src/handlers/poster/gewis/components/GewisProgressBar.tsx
@@ -1,5 +1,5 @@
-import ProgressBarSlider from '../../poster-base/components/ProgressBarSlider.tsx';
-import Clock from './Clock';
+import ProgressBarSlider from '../../components/ProgressBarSlider.tsx';
+import Clock from './Clock.tsx';
 
 interface Props {
   title?: string;
