+++ src/handlers/poster/gewis/components/types/BorrelLogoPoster.tsx
